Bind ToolWindow to per project object, not global ToolWindowFactory
